Don't use random magic number for id in tests
instead of using 1 or 2 we are now using list[0].id with list the list gotten in setup.
this prevent using an invalid id by error (which could break the controller/store.test)

also use -1 instead of 1000 for not found id to not cause a problem later if we have more
than 1000 value in a table.